refactor: extract `ConnectionState` into its own module
