+++ rust/libs/connlib/snownet/src/node.rs
@@ -1,4 +1,5 @@
 mod allocations;
+mod connection_state;
 mod connections;
 
 pub use connections::UnknownConnection;
@@ -6,6 +7,7 @@ pub use connections::UnknownConnection;
 use crate::allocation::{self, Allocation, RelaySocket, Socket};
 use crate::index::IndexLfsr;
 use crate::node::allocations::Allocations;
+use crate::node::connection_state::ConnectionState;
 use crate::node::connections::Connections;
 use crate::stats::{ConnectionStats, NodeStats};
 use crate::utils::channel_data_packet_buffer;
@@ -1294,236 +1296,6 @@ struct SelectedRelay<RId> {
     id: RId,
     /// Whether we've already logged failure to sample a new relay.
     logged_sample_failure: bool,
-}
-
-#[derive(Debug)]
-enum ConnectionState {
-    /// We are still running ICE to figure out, which socket to use to send data.
-    Connecting {
-        /// Packets emitted by wireguard whilst are still running ICE.
-        ///
-        /// This can happen if the remote's WG session initiation arrives at our socket before we nominate it.
-        /// A session initiation requires a response that we must not drop, otherwise the connection setup experiences unnecessary delays.
-        wg_buffer: AllocRingBuffer<Vec<u8>>,
-
-        /// Packets we are told to send whilst we are still running ICE.
-        ///
-        /// These need to be encrypted and sent once the tunnel is established.
-        ip_buffer: AllocRingBuffer<IpPacket>,
-    },
-    /// A socket has been nominated.
-    Connected {
-        /// Our nominated socket.
-        peer_socket: PeerSocket,
-
-        last_activity: Instant,
-    },
-    /// We haven't seen application packets in a while.
-    Idle {
-        /// Our nominated socket.
-        peer_socket: PeerSocket,
-    },
-    /// The connection failed in an unrecoverable way and will be GC'd.
-    Failed,
-}
-
-impl ConnectionState {
-    fn poll_timeout(&self, agent: &IceAgent) -> Option<(Instant, &'static str)> {
-        if agent.state() != IceConnectionState::Connected {
-            return None;
-        }
-
-        match self {
-            ConnectionState::Connected { last_activity, .. } => {
-                Some((idle_at(*last_activity), "idle transition"))
-            }
-            ConnectionState::Connecting { .. }
-            | ConnectionState::Idle { .. }
-            | ConnectionState::Failed => None,
-        }
-    }
-
-    fn handle_timeout(&mut self, agent: &mut IceAgent, idle_ice_config: IceConfig, now: Instant) {
-        let Self::Connected {
-            last_activity,
-            peer_socket,
-        } = self
-        else {
-            return;
-        };
-
-        if idle_at(*last_activity) > now {
-            return;
-        }
-
-        if agent.state() != IceConnectionState::Connected {
-            return;
-        }
-
-        let peer_socket = *peer_socket;
-
-        self.transition_to_idle(peer_socket, agent, idle_ice_config);
-    }
-
-    fn on_upsert<TId>(
-        &mut self,
-        cid: TId,
-        agent: &mut IceAgent,
-        default_ice_config: IceConfig,
-        now: Instant,
-    ) where
-        TId: fmt::Display,
-    {
-        let peer_socket = match self {
-            Self::Idle { peer_socket } => *peer_socket,
-            Self::Connected { last_activity, .. } => {
-                *last_activity = now;
-                return;
-            }
-            Self::Failed | Self::Connecting { .. } => return,
-        };
-
-        self.transition_to_connected(cid, peer_socket, agent, default_ice_config, "upsert", now);
-    }
-
-    fn on_candidate<TId>(
-        &mut self,
-        cid: TId,
-        agent: &mut IceAgent,
-        default_ice_config: IceConfig,
-        now: Instant,
-    ) where
-        TId: fmt::Display,
-    {
-        let peer_socket = match self {
-            Self::Idle { peer_socket } => *peer_socket,
-            Self::Connected { last_activity, .. } => {
-                *last_activity = now;
-                return;
-            }
-            Self::Failed | Self::Connecting { .. } => return,
-        };
-
-        self.transition_to_connected(
-            cid,
-            peer_socket,
-            agent,
-            default_ice_config,
-            "candidates changed",
-            now,
-        );
-    }
-
-    fn on_outgoing<TId>(
-        &mut self,
-        cid: TId,
-        agent: &mut IceAgent,
-        default_ice_config: IceConfig,
-        packet: &IpPacket,
-        now: Instant,
-    ) where
-        TId: fmt::Display,
-    {
-        let peer_socket = match self {
-            Self::Idle { peer_socket } => *peer_socket,
-            Self::Connected { last_activity, .. } => {
-                *last_activity = now;
-                return;
-            }
-            Self::Failed | Self::Connecting { .. } => return,
-        };
-
-        self.transition_to_connected(
-            cid,
-            peer_socket,
-            agent,
-            default_ice_config,
-            tracing::field::debug(packet),
-            now,
-        );
-    }
-
-    fn on_incoming<TId>(
-        &mut self,
-        cid: TId,
-        agent: &mut IceAgent,
-        default_ice_config: IceConfig,
-        packet: &IpPacket,
-        now: Instant,
-    ) where
-        TId: fmt::Display,
-    {
-        let peer_socket = match self {
-            Self::Idle { peer_socket } => *peer_socket,
-            Self::Connected { last_activity, .. } => {
-                *last_activity = now;
-                return;
-            }
-            Self::Failed | Self::Connecting { .. } => return,
-        };
-
-        self.transition_to_connected(
-            cid,
-            peer_socket,
-            agent,
-            default_ice_config,
-            tracing::field::debug(packet),
-            now,
-        );
-    }
-
-    fn transition_to_idle(
-        &mut self,
-        peer_socket: PeerSocket,
-        agent: &mut IceAgent,
-        idle_ice_config: IceConfig,
-    ) {
-        tracing::debug!("Connection is idle");
-        *self = Self::Idle { peer_socket };
-        idle_ice_config.apply(agent);
-    }
-
-    fn transition_to_connected<TId>(
-        &mut self,
-        cid: TId,
-        peer_socket: PeerSocket,
-        agent: &mut IceAgent,
-        default_ice_config: IceConfig,
-        trigger: impl tracing::Value,
-        now: Instant,
-    ) where
-        TId: fmt::Display,
-    {
-        tracing::debug!(trigger, %cid, "Connection resumed");
-        *self = Self::Connected {
-            peer_socket,
-            last_activity: now,
-        };
-        default_ice_config.apply(agent);
-    }
-
-    fn has_nominated_socket(&self) -> bool {
-        matches!(self, Self::Connected { .. } | Self::Idle { .. })
-    }
-}
-
-impl fmt::Display for ConnectionState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ConnectionState::Connecting { .. } => write!(f, "Connecting"),
-            ConnectionState::Connected { peer_socket, .. } => {
-                write!(f, "Connected({})", peer_socket.kind())
-            }
-            ConnectionState::Idle { peer_socket } => write!(f, "Idle({})", peer_socket.kind()),
-            ConnectionState::Failed => write!(f, "Failed"),
-        }
-    }
-}
-
-fn idle_at(last_activity: Instant) -> Instant {
-    const MAX_IDLE: Duration = Duration::from_secs(20); // Must be longer than the ICE timeout otherwise we might not detect a failed connection early enough.
-
-    last_activity + MAX_IDLE
 }
 
 /// The socket of the peer we are connected to.
